update jenkins cookbook

+++ cookbooks/jenkins/recipes/default.rb
@@ -7,30 +7,45 @@
 # All rights reserved - Do Not Redistribute
 #
 
+jenkins_home ||= "/var/lib/jenkins"
+plugin_path = "#{jenkins_home}/plugins"
+plugin_repo = "http://updates.jenkins-ci.org/latest/"
+plugins = ['gradle','git','greenballs']
+
 bash "setup_jenkins_repo" do
-  code <<-EOH
-    wget -O /etc/yum.repos.d/jenkins.repo http://pkg.jenkins-ci.org/redhat/jenkins.repo
-    rpm --import http://pkg.jenkins-ci.org/redhat/jenkins-ci.org.key
-  EOH
-  action :run
+  if ::File.exists?("/etc/yum.repos.d/jenkins.repo")
+    action :nothing
+  else
+    code <<-EOH
+      wget -O /etc/yum.repos.d/jenkins.repo http://pkg.jenkins-ci.org/redhat/jenkins.repo
+      rpm --import http://pkg.jenkins-ci.org/redhat/jenkins-ci.org.key
+    EOH
+    action :run
+  end
 end
 
 yum_package "jenkins" do
-  action :install
+  if `rpm -qa | grep jenkins`.empty?
+    action :install
+  else
+    action :nothing
+  end
 end
 
-bash "download_jenkins_plugin_gradle" do
-  code <<-EOH
-    wget -O /var/lib/jenkins/plugins/gradle.hpi http://updates.jenkins-ci.org/latest/gradle.hpi --no-check-certificate
-  EOH
-  action :run
-end
+#bash "download_jenkins_plugin_gradle" do
+#  code <<-EOH
+#    wget -O /var/lib/jenkins/plugins/gradle.hpi http://updates.jenkins-ci.org/latest/gradle.hpi --no-check-certificate
+#  EOH
+#  action :run
+#end
 
+plugins.each do |plugin|
+  local="#{plugin_path}/#{plugin}.hpi"
+  remote="#{plugin_repo}/#{plugin}.hpi"
 
-directory "/tmp/chef/test" do
-  owner "root"
-  group "root"
-  mode 00644
-  recursive true
-  action :create
+  remote_file local do
+    source remote
+    mode "0644"
+    not_if { ::File.exists?(local) }
+  end
 end
